[👕] absolute k indexing issues are fixed
What was written in these two lines is working now that absolute k
indexing is (as experimental feature) in mainline gt4py. PR
https://github.com/NOAA-GFDL/NDSL/pull/303 merged a version of gt4py
with explicit test coverage for the case of computations in index
access. The PR also brought a fix up from gt4py where computations with
different types weren't possible on the left hand side of an assignment.